refactor; utils <- header-only
