+++ src/par_huffman.cu
@@ -20,14 +20,14 @@
 #include <cstring>
 #include <limits>
 
-#include "cuda_error_handling.cuh"
-#include "cuda_mem.cuh"
-#include "dbg_gpu_printing.cuh"
-#include "format.hh"
 #include "par_huffman.cuh"
 #include "par_huffman_sortbyfreq.cuh"
 #include "par_merge.cuh"
 #include "type_aliasing.hh"
+#include "utils/cuda_err.cuh"
+#include "utils/cuda_mem.cuh"
+#include "utils/dbg_print.cuh"
+#include "utils/format.hh"
 
 // Mathematically correct mod
 #define MOD(a, b) ((((a) % (b)) + (b)) % (b))
